requirements.txt

# General dependencies
requests

# Google dependencies
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
markdown2

# CrewAI dependencies
crewai
crewai_tools

# OpenAI dependencies
openai
